Avoid shallow clone in github-actions example
A shallow clone will not have the history we need in order to create the
base branch. Using fetch-depth:0 would also work (it fetches the entire
history) but that would make things really slow in some cases. So I
picked 100 as a good-enough number.

Please note that this change is untested. I used the docs here:
https://github.com/actions/checkout#fetch-all-history-for-all-tags-and-branches