Add documentation for stabilizing stack traces in spurious diffs
I noticed that some customers were seeing spurious diffs due to stack
traces changing all of the time. I'd like to send them some advice on
how to fix this, so I figured this is a good opportunity to improve our
documentation.

To make this document easier to expand, I changed the bulleted list to a
section with headings. This will make it easier for us to add more code
examples in the future.