Recommend npx in CI documentation
I think this will be a more foolproof way to get Happo running in CI.
And as a result of this we can simplify our instructions a bit.

I also re-ordered the sections based on what I think might be the most
popular CI providers (GitHub Actions, CircleCI, Azure, Travis, and then
generic).

Along the way I updated a couple of things that were outdated.